Move lib files from /lib/tasks into /lib
- move corresponding test files
