+++ lib/tasks/sanitize_data.rake
@@ -1,11 +1,11 @@
 desc "Sanitize access limited data"
 task sanitize_data: :environment do
-  Tasks::DataSanitizer.delete_access_limited($stdout)
+  DataSanitizer.delete_access_limited($stdout)
 end
 
 namespace :db do
   desc "Resolves invalid versions detected by validate:versions task"
   task resolve_invalid_versions: :environment do
-    Tasks::VersionResolver.resolve
+    VersionResolver.resolve
   end
 end
