+++ lib/tasks/validate.rake
@@ -1,11 +1,11 @@
 namespace :db do
   desc "Validates every record in the database"
   task validate: :environment do
-    Tasks::DatabaseRecordValidator.validate
+    DatabaseRecordValidator.validate
   end
 
   desc "Validates the version sequence for all editions in the database"
   task validate_versions: :environment do
-    Tasks::VersionValidator.validate
+    VersionValidator.validate
   end
 end
